pyproject.toml
[tool.poetry]
name = "medagogic-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "packages/server" },
    { include = "packages/tools" },
    { include = "packages/medagogic_sim" }
  ]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
annotated-types = "0.5.0"
anyio = "3.7.1"
async-timeout = "4.0.2"
attrs = "23.1.0"
backoff = "2.2.1"
bcrypt = "4.0.1"
bidict = "0.22.1"
boto3 = "1.28.74"
boto3-stubs = "1.28.74"
botocore = "1.31.74"
botocore-stubs = "1.31.74"
cachetools = "5.3.1"
certifi = "2023.5.7"
cffi = "1.16.0"
charset-normalizer = "3.1.0"
chroma-hnswlib = "0.7.3"
chromadb = "0.4.14"
click = "8.1.7"
colorama = "0.4.6"
coloredlogs = "15.0.1"
contourpy = "1.1.1"
cryptography = "41.0.4"
cycler = "0.12.1"
docker = "6.1.3"
docstring-parser = "0.15"
exceptiongroup = "1.1.3"
faiss-cpu = "1.7.4"
fastapi = "0.103.1"
fastapi-socketio = "0.0.10"
filelock = "3.12.4"
flatbuffers = "23.5.26"
fonttools = "4.43.1"
frozenlist = "1.3.3"
fsspec = "2023.10.0"
google-api-core = "2.12.0"
google-auth = "2.23.3"
google-cloud-appengine-logging = "1.3.2"
google-cloud-audit-log = "0.2.5"
google-cloud-core = "2.3.3"
google-cloud-logging = "3.8.0"
googleapis-common-protos = "1.61.0"
gptcache = "0.1.42"
greenlet = "3.0.0"
grpc-google-iam-v1 = "0.12.6"
grpcio = "1.59.0"
grpcio-status = "1.59.0"
h11 = "0.14.0"
hiredis = "2.2.3"
httpcore = "0.18.0"
httptools = "0.6.1"
httpx = "0.25.0"
huggingface-hub = "0.17.3"
human-id = "0.2.0"
humanfriendly = "10.0"
idna = "3.4"
importlib-resources = "6.1.0"
iniconfig = "2.0.0"
jinja2 = "3.1.2"
jmespath = "1.0.1"
joblib = "1.3.2"
jsonschema = "4.19.1"
jsonschema-specifications = "2023.7.1"
jsonschema2md = "1.0.0"
kiwisolver = "1.4.5"
markdown-to-json = "2.1.0"
markupsafe = "2.1.3"
matplotlib = "3.8.0"
monotonic = "1.6"
more-itertools = "9.1.0"
moto = "4.2.7"
mpmath = "1.3.0"
multidict = "6.0.4"
mypy = "1.5.1"
mypy-boto3-dynamodb = "1.28.73"
mypy-extensions = "1.0.0"
nltk = "3.8.1"
onnxruntime = "1.16.1"
openai = "0.27.7"
overrides = "7.4.0"
packaging = "23.2"
pandas = "2.1.2"
pillow = "10.1.0"
plotly = "5.18.0"
pluggy = "1.3.0"
posthog = "3.0.2"
prompt-toolkit = "3.0.39"
proto-plus = "1.22.3"
protobuf = "4.24.4"
pulsar-client = "3.3.0"
py-jama-rest-client = "1.17.1"
py-partiql-parser = "0.4.1"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.0.3"
pydantic-core = "2.3.0"
pydeps = "1.12.17"
pyparsing = "3.1.1"
pypika = "0.48.9"
pysqlite3-binary = "0.5.2.post1"
pytest = "7.4.2"
pytest-asyncio = "0.21.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-engineio = "4.7.1"
python-socketio = "5.9.0"
python-ulid = "1.1.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
redis = "4.6.0"
redis-om = "0.2.1"
referencing = "0.30.2"
regex = "2023.10.3"
requests = "2.31.0"
responses = "0.23.3"
rpds-py = "0.10.6"
rsa = "4.9"
rx = "3.2.0"
s3transfer = "0.7.0"
scikit-learn = "1.3.2"
simple-websocket = "0.10.1"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.22"
starlette = "0.27.0"
stdlib-list = "0.9.0"
sympy = "1.12"
tenacity = "8.2.2"
threadpoolctl = "3.2.0"
tiktoken = "0.5.1"
tokenizers = "0.14.1"
tomli = "2.0.1"
tqdm = "4.65.0"
typer = "0.9.0"
types-awscrt = "0.19.7"
types-pyopenssl = "23.2.0.2"
types-pyyaml = "6.0.12.12"
types-redis = "4.6.0.7"
types-s3transfer = "0.7.0"
typing-extensions = "4.8.0"
tzdata = "2023.3"
urllib3 = "1.26.18"
uvicorn = "0.23.2"
uvloop = "0.19.0"
watchfiles = "0.21.0"
wcwidth = "0.2.8"
websocket-client = "1.6.4"
websockets = "12.0"
werkzeug = "3.0.1"
wsproto = "1.2.0"
xmltodict = "0.13.0"
yarl = "1.9.2"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
